Fetching environments in backend
