+++ app-config.yaml
@@ -89,3 +89,4 @@ catalog:
 humanitec:
   orgId: the-frontside-software-inc
   registryUrl: "northamerica-northeast1-docker.pkg.dev/frontside-backstage/frontside-artifacts"
+  "proxyPath": "/humanitec"
