+++ plugins/humanitec-backend/src/schemas/deployment.ts
@@ -1,0 +1,3 @@
+import { object } from 'zod';
+
+export const Deployment = object({});
